Change findTest to return the most recently added test
Whenever we encounter a result line in the test output, we would find
the first test with the matching name and update it. However, in some
cases it's possible for the same test name to appear multiple times in
the output. To prevent us from always updating the oldest test, the
order in which findTests searches for a matching test is reversed so
that it always returns the most recently added test.

Fixes #54.