Deprecate the `-go-version` flag
The `-go-version` flag is used to set the `go.version` property in the
generated XML file. In v2 we no longer want a flag dedicated to a
specific property. Instead, the -prop flag has been introduced to set
arbitrary properties in the generated report.

If the `-go-version` flag is set, we'll still add a `go.version`
property to the report but also print a warning. This flag will be
removed completely in the future.