Refactor and simplify tests
For most testcases we shouldn't need to specify the input and report
files, instead just look at the testdata/ dir and process each input
file. For the few testcases that need a non-standard config, an entry
can be added to `testConfigs` map (indexed by the test number).

The testdata files must follow a certain naming pattern to be valid.
They should start with a number, followed by a dash, then any text
describing the test and finally end with `.txt`. The corresponding
report file must start with the same number, a dash, followed by
`report.xml`.