Regenerate testdata files
Many of the testdata files were generated a long time ago with very old
versions of Go. The Go test output has changed over time, and these test
inputs no longer reflect the current state. The intention is to fully
support the test output of the most recent Go versions.

This commit also includes the source used to generate the test output,
so the output can be more easily updated in the future.